Replace solenoid side with similar attachment as spring
I realised that the sliding-pin technique was too complex for what it needed to do and would introduce much more friction than necessary. Instead, the assembly will now require an additional part: A linear bar to transfer force from the solenoid to the grip. But this allows it to rotate the attachment between that bar and the grip around just a single point instead of having a sliding pin sort of construction.