Add waves sprites
These waves will be displayed in the score screen.
The waves have different sizes. There's 4 sizes, one for each height of pixel (up to 4 pixels high). For each size there are 8 sprites, all aligned to a multiple of 4 pixel rows.

It is intended to process these sprites with a Python script to data arrays that will be included in the source.