Add Python script to pack wave sprite data into source
This converts the sprite into C source code that can be included to have the pixel data in a variable.
This is necessary because this device will not have any flash drive or file system or anything. So you can't load this data from file, so it has to be included statically in the source code.