Generate one monochrome image per colour
The default draw routine only supports monochrome images. So we have to reduce the colour depth to 8 colours and then split up each image into 8 monochrome images, of which we leave out the 0-colour (interpreting it as transparent). The waves.h header becomes less readable, but at least it should work (not tested yet).