Add design for plunger
This is a metal rod with a spring. The rod is attached to a handle (with glue? screws?) and hanging loose on the other side. The ball rolls into the cavity at the loose end of the rod. The user can pull on the handle, compressing the spring, and then let go at some point, launching the rod forward just beyond the gate, slamming against the ball and thus launching the ball into the playfield.